# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Steam4Caster API",
    "description": "\nBackend API for **Steam4Caster**: probabilistic Steam discount forecasts, explainable\nBUY / WAIT / NEUTRAL recommendations, watchlists and alerts.\n\n* All public routes are versioned under `/api/v1`.\n* Authenticate with `Authorization: Bearer <access token>`.\n* Money is always integer minor units plus an ISO 4217 currency, per region. Prices are\n  never converted between currencies.\n* Errors use one envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\n* Forecasts and recommendations are statistical estimates based on past prices. They are not guarantees and not financial advice.\n\nPrice data is provided by the IsThereAnyDeal API. Steam4Caster is not affiliated with\nIsThereAnyDeal or Valve.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/model-performance/summary": {
      "get": {
        "tags": [
          "model-performance"
        ],
        "summary": "Aggregate accuracy of matured forecasts",
        "operationId": "performance_summary_api_v1_model_performance_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one model version.",
              "title": "Model Version"
            },
            "description": "Restrict to one model version."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceSummary"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/model-performance/calibration": {
      "get": {
        "tags": [
          "model-performance"
        ],
        "summary": "Reliability bins: predicted probability vs observed sale rate",
        "operationId": "performance_calibration_api_v1_model_performance_calibration_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "One of 7, 30 or 90.",
              "default": 30,
              "title": "Horizon Days"
            },
            "description": "One of 7, 30 or 90."
          },
          {
            "name": "bins",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 2,
              "default": 10,
              "title": "Bins"
            }
          },
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one model version.",
              "title": "Model Version"
            },
            "description": "Restrict to one model version."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/model-performance/by-horizon": {
      "get": {
        "tags": [
          "model-performance"
        ],
        "summary": "Brier score, log loss and ranking metrics per horizon",
        "operationId": "performance_by_horizon_api_v1_model_performance_by_horizon_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one model version.",
              "title": "Model Version"
            },
            "description": "Restrict to one model version."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ByHorizonOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/model-versions/active": {
      "get": {
        "tags": [
          "model-performance"
        ],
        "summary": "The model version currently used for forecasts",
        "operationId": "active_model_api_v1_model_versions_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveModelOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/meta": {
      "get": {
        "tags": [
          "model-performance",
          "system"
        ],
        "summary": "Public client configuration: attribution, VAPID key, tiers and horizons",
        "operationId": "meta_api_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActiveModelOut": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelVersionOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when no ML model is active and the deterministic baseline is in use."
          },
          "fallback_method": {
            "type": "string",
            "title": "Fallback Method",
            "default": "BASELINE"
          },
          "baseline_version": {
            "type": "string",
            "title": "Baseline Version"
          },
          "ruleset_version": {
            "type": "string",
            "title": "Ruleset Version"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version"
          }
        },
        "type": "object",
        "required": [
          "active",
          "baseline_version",
          "ruleset_version",
          "feature_schema_version"
        ],
        "title": "ActiveModelOut"
      },
      "Attribution": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "affiliation": {
            "type": "string",
            "title": "Affiliation"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "text",
          "url",
          "affiliation"
        ],
        "title": "Attribution"
      },
...
      "ByHorizonOut": {
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "horizons": {
            "items": {
              "$ref": "#/components/schemas/HorizonMetrics"
            },
            "type": "array",
            "title": "Horizons"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "horizons"
        ],
        "title": "ByHorizonOut"
      },
      "CalibrationBin": {
        "properties": {
          "bin_lower": {
            "type": "number",
            "title": "Bin Lower"
          },
          "bin_upper": {
            "type": "number",
            "title": "Bin Upper"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "mean_predicted": {
            "type": "number",
            "title": "Mean Predicted"
          },
          "observed_rate": {
            "type": "number",
            "title": "Observed Rate"
          }
        },
        "type": "object",
        "required": [
          "bin_lower",
          "bin_upper",
          "count",
          "mean_predicted",
          "observed_rate"
        ],
        "title": "CalibrationBin"
      },
      "CalibrationOut": {
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "bins": {
            "items": {
              "$ref": "#/components/schemas/CalibrationBin"
            },
            "type": "array",
            "title": "Bins"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "horizon_days",
          "n",
          "bins"
        ],
        "title": "CalibrationOut"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HorizonMetrics": {
        "properties": {
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "positive_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Positive Rate"
          },
          "brier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brier"
          },
          "log_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Loss"
          },
          "ece": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ece",
            "description": "Expected calibration error."
          },
          "roc_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roc Auc"
          },
          "pr_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Auc"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "precision_at_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision At Threshold"
          },
          "recall_at_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall At Threshold"
          }
        },
        "type": "object",
        "required": [
          "horizon_days",
          "n"
        ],
        "title": "HorizonMetrics"
      },
...
      "MetaOut": {
        "properties": {
          "app_name": {
            "type": "string",
            "title": "App Name"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "attribution": {
            "$ref": "#/components/schemas/Attribution"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          },
          "default_country": {
            "type": "string",
            "title": "Default Country"
          },
          "default_currency": {
            "type": "string",
            "title": "Default Currency"
          },
          "supported_shops": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Shops"
          },
          "forecast_horizons_days": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Forecast Horizons Days"
          },
          "discount_tiers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Discount Tiers"
          },
          "channels": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Channels"
          },
          "vapid_public_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vapid Public Key"
          },
          "ruleset_version": {
            "type": "string",
            "title": "Ruleset Version"
          },
          "baseline_version": {
            "type": "string",
            "title": "Baseline Version"
          }
        },
        "type": "object",
        "required": [
          "app_name",
          "api_version",
          "attribution",
          "disclaimer",
          "default_country",
          "default_currency",
          "supported_shops",
          "forecast_horizons_days",
          "discount_tiers",
          "channels",
          "vapid_public_key",
          "ruleset_version",
          "baseline_version"
        ],
        "title": "MetaOut"
      },
      "ModelVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "training_cutoff": {
            "type": "string",
            "format": "date-time",
            "title": "Training Cutoff"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version"
          },
          "hyperparameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hyperparameters"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "status",
          "training_cutoff",
          "feature_schema_version",
          "hyperparameters",
          "metrics",
          "created_at",
          "activated_at"
        ],
        "title": "ModelVersionOut"
      },
...
      "PerformanceSummary": {
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "forecasts_total": {
            "type": "integer",
            "title": "Forecasts Total"
          },
          "forecasts_evaluated": {
            "type": "integer",
            "title": "Forecasts Evaluated"
          },
          "sale_probability": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Sale Probability"
          },
          "by_method": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "By Method"
          },
          "discount_tier": {
            "additionalProperties": true,
            "type": "object",
            "title": "Discount Tier"
          },
          "recommendation_policy": {
            "additionalProperties": true,
            "type": "object",
            "title": "Recommendation Policy"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "forecasts_total",
          "forecasts_evaluated",
          "sale_probability",
          "by_method",
          "discount_tier",
          "recommendation_policy",
          "disclaimer"
        ],
        "title": "PerformanceSummary"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/v1/auth/login.",
        "scheme": "bearer"
      }
    }
  }
}